-Added time step forward and backward functionality in the editing window controller class. -As a whole, the code requires the user to enter in information for chick 1-3 in the order for every operation, whether its adding new information or modifying old data -The markers now display if the user goes over a frame they have already collecting data on -Added some methods in the animalTrack class to colect information from a given animalTrack object

+++ ProjectEgret/src/edu/augustana/csc285/Egret/MainWindowController.java
@@ -74,7 +74,7 @@ public class MainWindowController {
 	
 	private void startVideo() {
 		vidCap.open(fileName);
-		video.numFrame = vidCap.get(Videoio.CV_CAP_PROP_FRAME_COUNT);
+		video.numFrame = (int) vidCap.get(Videoio.CV_CAP_PROP_FRAME_COUNT);
 		totalFrameArea.appendText("Total Frames: " + (int) video.numFrame + "\n");
 		slider.setDisable(false);
 		jumpToFrameArea.setDisable(false);
